feat: Add GitHub Actions workflow for linting code
Add a new GitHub Actions workflow file, super-linter.yml, to the repository. This workflow is responsible for running the super-linter tool to perform code linting. It is triggered on push and pull request events. The workflow runs on the latest version of Ubuntu and has the necessary permissions to read the repository contents and packages, as well as write statuses for GitHub Actions status checks. The workflow checks out the code and uses the super-linter action to perform linting. The GITHUB_TOKEN secret is used for reporting GitHub Actions status checks.